backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
base58==2.1.1
certifi==2025.1.31
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
ECPy==1.2.5
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.7.0
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pydantic-settings==2.8.1
Pygments==2.19.1
pymongo==4.12.0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typer==0.15.2
types-Deprecated==1.2.15.20250304
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
xrpl-py==4.1.0
charset-normalizer==3.4.1                                                                                                                     
jsonpatch==1.33                                                                                                                               
jsonpointer==3.0.0                                                                                                                            
langchain==0.3.23                                                                                                                             
langchain-core==0.3.51                                                                                                                        
langchain-text-splitters==0.3.8                                                                                                               
langsmith==0.3.30                                                                                                                             
orjson==3.10.16                                                                                                                               
packaging==24.2                                                                                                                               
requests==2.32.3                                                                                                                              
requests-toolbelt==1.0.0                                                                                                                      
urllib3==2.4.0                                                                                                                                
zstandard==0.23.0
langchain-openai==0.3.12                                                                                                                      
regex==2024.11.6
tiktoken==0.9.0